Move merge code into its own class
To help manage complexity, move the merger code into its own class.
No code changes here.
