+++ Makefile
@@ -42,6 +42,7 @@ tmpl_nonexe_files = \
 
 util_nonexe_files = \
 	util/__init__.py \
+	util/debfilemerger.py \
 	util/jinja2-AUTHORS \
 	util/jinja.py \
 	util/shell.py \
